separated generateDeclaration function from initDb, and moved all db related tools to database/management folder
