+++ database/management/init-db.js
@@ -4,34 +4,12 @@
 
 require('es6-shim');
 
-var fs = require('fs');
-var path = require('path');
-
-var generateSqlDefinition = require('sql-generate');
-
-var PRIVATE = require('../PRIVATE/secret.json');
-
 // when ready, drop and create tables
-var databaseClientP = require('../database/management/databaseClientP');
-var dropAllTables = require('../database/management/dropAllTables.js');
-var createTables = require('../database/management/createTables.js');
+var databaseClientP = require('./databaseClientP');
+var dropAllTables = require('./dropAllTables.js');
+var createTables = require('./createTables.js');
 
-var conString = 'postgres://'+ PRIVATE.pg_user + ':' + PRIVATE.pg_pwd + '@localhost:5432/' + PRIVATE.db_name;
-
-console.log('Init-db connection string', conString);
-
-function generateDefinitions() {
-    return new Promise(function(resolve, reject) {
-        generateSqlDefinition({ dsn: conString, omitComments: true }, function(err, definitions) {
-            if (err) {
-                console.error(err);
-                reject(err);
-            }
-            fs.writeFileSync(path.join(__dirname, '../database/management/declarations.js'), definitions.buffer);
-            resolve();
-        });
-    });
-}
+var generateDeclarations = require('./generateDeclarations.js');
 
 (function tryRebuildDatabase(){
     console.log('Trying to rebuild database...');
@@ -51,7 +29,7 @@ function generateDefinitions() {
             })
             .then(function(){
                
-                generateDefinitions()
+                generateDeclarations()
                 .then(function(){
                     console.log('Dropped and created the tables.');
                     process.exit();
